refactor: extract commands from main
